Move common code to UnitTests.Mocks
