+++ src/UnitTests/Analysis/CoalescerTests.cs
@@ -28,8 +28,6 @@ using System;
 using System.IO;
 using System.Collections.Generic;
 using Reko.Core.Expressions;
-using Reko.Core.Output;
-using System.Diagnostics;
 
 namespace Reko.UnitTests.Analysis
 {
@@ -51,24 +49,9 @@ namespace Reko.UnitTests.Analysis
             m.Ssa.CheckUses(s => Assert.Fail(s));
         }
 
-        //$REFACTOR: move common code to Reko.UnitTests.Mocks
         private void AssertProcedureCode(string expected)
         {
-            var writer = new StringWriter();
-            var textFormatter = new TextFormatter(writer)
-            {
-                Indentation = 0,
-            };
-            textFormatter.WriteLine();
-            var codeFormatter = new CodeFormatter(textFormatter);
-            foreach (var stm in m.Ssa.Procedure.Statements)
-                stm.Instruction.Accept(codeFormatter);
-            var actual = writer.ToString();
-            if (expected != actual)
-            {
-                Debug.Print(actual);
-                Assert.AreEqual(expected, actual);
-            }
+            ProcedureCodeVerifier.AssertCode(m.Ssa.Procedure, expected);
         }
 
 		protected override void RunTest(Program program, TextWriter fut)
