+++ src/UnitTests/UnitTests.csproj
@@ -203,6 +203,8 @@
     <Compile Include="ImageLoaders\MzExe\Borland\SymbolLoaderTests.cs" />
     <Compile Include="Mocks\FakeTypeDeserializer.cs" />
     <Compile Include="Core\Analysis\PyBuildValueFormatParserTests.cs" />
+    <Compile Include="Mocks\ProcedureCodeVerifier.cs" />
+    <Compile Include="Mocks\ProcedureStatementsWriter.cs" />
     <Compile Include="Mocks\SsaProcedureBuilder.cs" />
     <Compile Include="Mocks\XmlnsHidingWriter.cs" />
     <Compile Include="OsPath.cs" />
